Add a replacement for prediction removal
Hello, to alter weapon fire sound we have to remove weapon prediction, but this also results in removal of bullet holes with smokes and weapon fire animation, beyond of slightly increase ping of players.

I believe there's a more appropiate way of do this, that would be cancel the call of PLAYBACK_EVENT_FULL, although it also results in removal of bullet holes and probably weapon fire animation.

We could add a member "m_bCantPredict" or something like. I did this branch for test purpose. If the method work, I'll proceed with the addition of this member.